requirements.txt
google-genai
pillow
gTTS
pandas
pyyaml
tqdm

# Hugging Face dependencies
transformers>=4.40.0
torch>=2.0.0
accelerate>=0.20.0
bitsandbytes>=0.41.0  # For 4-bit quantization (optional, GPU only)
sentencepiece  # For some tokenizers
protobuf  # For some models